Move pipe into shared module
