+++ CRBM-Viz/src/app/Layout/navigation/navigation.component.spec.ts
@@ -10,7 +10,7 @@ import { FooterComponent } from '../footer/footer.component';
 import { RouterTestingModule } from '@angular/router/testing';
 import { MaterialModule } from 'src/app/Modules/app-material.module';
 import { AccountMenuComponent } from '../account-menu/account-menu.component';
-import { FilterPipe } from 'src/app/Pipes/filter.pipe';
+import { FilterPipe } from 'src/app/Shared/Pipes/filter.pipe';
 import { SidebarComponent } from '../sidebar/sidebar.component';
 // TODO Sample content is needed and component compiled with it as input for <ng-content>
 import { AboutComponent } from 'src/app/Pages/about/about.component';
