+++ CRBM-Viz/src/app/Layout/sidebar/sidebar.component.spec.ts
@@ -3,7 +3,7 @@ import { SidebarComponent } from './sidebar.component';
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { RouterTestingModule } from '@angular/router/testing';
-import { FilterPipe } from 'src/app/Pipes/filter.pipe';
+import { FilterPipe } from 'src/app/Shared/Pipes/filter.pipe';
 describe('SidebarComponent', () => {
   let component: SidebarComponent;
   let fixture: ComponentFixture<SidebarComponent>;
