+++ CRBM-Viz/src/app/Shared/shared.module.ts
@@ -1,0 +1,8 @@
+import { NgModule } from '@angular/core';
+import { FilterPipe } from './Pipes/filter.pipe';
+
+@NgModule({
+  declarations: [FilterPipe],
+  exports: [FilterPipe],
+})
+export class SharedModule {}
